+++ src/test/java/net/sf/jabref/logic/formatter/FormatterTest.java
@@ -2,6 +2,7 @@ package net.sf.jabref.logic.formatter;
 
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.logic.formatter.bibtexfields.ClearFormatter;
@@ -24,6 +25,8 @@ import net.sf.jabref.logic.formatter.casechanger.TitleCaseFormatter;
 import net.sf.jabref.logic.formatter.casechanger.UpperCaseFormatter;
 import net.sf.jabref.logic.formatter.minifier.MinifyNameListFormatter;
 import net.sf.jabref.logic.layout.format.LatexToUnicodeFormatter;
+import net.sf.jabref.logic.protectedterms.ProtectedTermsLoader;
+import net.sf.jabref.logic.protectedterms.ProtectedTermsPreferences;
 import net.sf.jabref.preferences.JabRefPreferences;
 
 import org.junit.BeforeClass;
@@ -49,6 +52,10 @@ public class FormatterTest {
     @BeforeClass
     public static void setUp() {
         Globals.prefs = JabRefPreferences.getInstance();
+        ProtectTermsFormatter
+                .setProtectedTermsLoader(
+                        new ProtectedTermsLoader(new ProtectedTermsPreferences(ProtectedTermsLoader.getInternalLists(),
+                                Collections.emptyList(), Collections.emptyList(), Collections.emptyList())));
     }
 
     @Test